Moved HandleError to shared
